limit test to wasm
