+++ src/bytes/bytes_test.go
@@ -2278,12 +2278,3 @@ func TestClone(t *testing.T) {
 		}
 	}
 }
-
-func TestIssue65571(t *testing.T) {
-	b := make([]byte, 1<<31+1)
-	b[1<<31] = 1
-	i := IndexByte(b, 1)
-	if i != 1<<31 {
-		t.Errorf("IndexByte(b, 1) = %d; want %d", i, 1<<31)
-	}
-}
